requirements.txt
pandas
scikit-learn
torch
kafka-python
